refactor: extract input modal helpers
